Deleting the helper class
