+++ src/commands/deployment/get.ts
@@ -1,10 +1,29 @@
+import Table from "cli-table";
 import commander from "commander";
+import Deployment from "spektate/lib/Deployment";
 import { logger } from "../../logger";
-import {
-  getDeployments,
-  OUTPUT_FORMAT,
-  verifyAppConfiguration
-} from "./helper";
+import { verifyAppConfiguration } from "./init";
+import { clusterPipeline, config, hldPipeline, srcPipeline } from "./init";
+
+/**
+ * Output formats to display service details
+ */
+export enum OUTPUT_FORMAT {
+  /**
+   * Normal format
+   */
+  NORMAL = 0,
+
+  /**
+   * Wide table format
+   */
+  WIDE = 1,
+
+  /**
+   * JSON format
+   */
+  JSON = 2
+}
 
 /**
  * Adds the get command to the commander command object
@@ -78,3 +97,138 @@ function processOutputFormat(outputFormat: string): OUTPUT_FORMAT {
 
   return OUTPUT_FORMAT.NORMAL;
 }
+
+/**
+ * Gets a list of deployments for the specified filters
+ */
+export const getDeployments = (
+  outputFormat: OUTPUT_FORMAT,
+  environment?: string,
+  imageTag?: string,
+  p1Id?: string,
+  commitId?: string,
+  service?: string,
+  deploymentId?: string
+): Promise<Deployment[]> => {
+  return Deployment.getDeploymentsBasedOnFilters(
+    config.STORAGE_ACCOUNT_NAME,
+    config.STORAGE_ACCOUNT_KEY,
+    config.STORAGE_TABLE_NAME,
+    config.STORAGE_PARTITION_KEY,
+    srcPipeline,
+    hldPipeline,
+    clusterPipeline,
+    environment,
+    imageTag,
+    p1Id,
+    commitId,
+    service,
+    deploymentId
+  ).then((deployments: Deployment[]) => {
+    if (outputFormat === OUTPUT_FORMAT.JSON) {
+      logger.info(JSON.stringify(deployments, null, 2));
+    } else {
+      printDeployments(deployments, outputFormat);
+    }
+    return deployments;
+  });
+};
+
+/**
+ * Prints deployments in a terminal table
+ */
+export const printDeployments = (
+  deployments: Deployment[],
+  outputFormat: OUTPUT_FORMAT
+) => {
+  if (deployments.length > 0) {
+    let row = [];
+    row.push("Start Time");
+    row.push("Service");
+    row.push("Deployment");
+    row.push("Commit");
+    row.push("Src to ACR");
+    row.push("Image Tag");
+    row.push("Result");
+    row.push("ACR to HLD");
+    row.push("Env");
+    row.push("Hld Commit");
+    row.push("Result");
+    row.push("HLD to Manifest");
+    row.push("Result");
+    if (outputFormat === OUTPUT_FORMAT.WIDE) {
+      row.push("Duration");
+      row.push("Status");
+      row.push("Manifest Commit");
+      row.push("End Time");
+    }
+    const table = new Table({ head: row });
+    deployments.forEach(deployment => {
+      row = [];
+      row.push(
+        deployment.srcToDockerBuild
+          ? deployment.srcToDockerBuild.startTime.toLocaleString()
+          : ""
+      );
+      row.push(deployment.service);
+      row.push(deployment.deploymentId);
+      row.push(deployment.commitId);
+      row.push(
+        deployment.srcToDockerBuild ? deployment.srcToDockerBuild.id : ""
+      );
+      row.push(deployment.imageTag);
+      row.push(
+        deployment.srcToDockerBuild
+          ? getStatus(deployment.srcToDockerBuild.result)
+          : ""
+      );
+      row.push(
+        deployment.dockerToHldRelease ? deployment.dockerToHldRelease.id : ""
+      );
+      row.push(deployment.environment.toUpperCase());
+      row.push(deployment.hldCommitId);
+      row.push(
+        deployment.dockerToHldRelease
+          ? getStatus(deployment.dockerToHldRelease.status)
+          : ""
+      );
+      row.push(
+        deployment.hldToManifestBuild ? deployment.hldToManifestBuild.id : ""
+      );
+      row.push(
+        deployment.hldToManifestBuild
+          ? getStatus(deployment.hldToManifestBuild.result)
+          : ""
+      );
+      if (outputFormat === OUTPUT_FORMAT.WIDE) {
+        row.push(deployment.duration() + " mins");
+        row.push(deployment.status());
+        row.push(deployment.manifestCommitId);
+        row.push(
+          deployment.hldToManifestBuild &&
+            deployment.hldToManifestBuild.finishTime &&
+            !isNaN(deployment.hldToManifestBuild.finishTime.getTime())
+            ? deployment.hldToManifestBuild.finishTime.toLocaleString()
+            : ""
+        );
+      }
+      table.push(row);
+    });
+
+    logger.info(table.toString());
+  } else {
+    logger.info("No deployments found for specified filters.");
+  }
+};
+
+/**
+ * Gets a status indicator icon
+ */
+const getStatus = (status: string) => {
+  if (status === "succeeded") {
+    return "\u2713";
+  } else if (!status) {
+    return "...";
+  }
+  return "\u0445";
+};
